feat: add git to the ncs-zephyr CI image
This fixes an issue in situation where the image is used to build a
container inside the GitHub Actions.

For some reason the git from the container is not being used in the
action/checkout, it always report that it doesn't find it on the PATH.
Although it can fallback to using an older version by downloading it
through the GitHub API, this doesn't work reliably.