feat: use newer version of the JLink
Version provided with the nrf commandline tools is quite old and is
incompatible with the XIP features.

This commit downloads a newer version and installs it to the system
path.

I decided to make this part of the base image (inherited in Dev and in
CI) and not only Dev, as CI already install nrfutil device command
depends on JLink and it would be weird if some workflow would start
to use nrfutil device and that wouldn't work as expected due to
missing JLink